packages/modules/AdServices/adservices/tests/unittest/system-service/src/com/android/server/adservices/rollback/RollbackHandlingManagerTest.java
/*
 * Copyright (C) 2023 The Android Open Source Project
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *      http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

package com.android.server.adservices.rollback;

import static com.android.dx.mockito.inline.extended.ExtendedMockito.any;
import static com.android.server.adservices.rollback.RollbackHandlingManager.STORAGE_XML_IDENTIFIER;
import static com.android.server.adservices.rollback.RollbackHandlingManager.VERSION_KEY;

import static com.google.common.truth.Truth.assertThat;

import static org.junit.Assert.assertThrows;

import android.app.adservices.AdServicesManager;
import android.content.Context;

import androidx.test.core.app.ApplicationProvider;

import com.android.dx.mockito.inline.extended.ExtendedMockito;
import com.android.server.adservices.common.BooleanFileDatastore;

import org.junit.After;
import org.junit.Before;
import org.junit.Test;
import org.mockito.MockitoSession;
import org.mockito.quality.Strictness;

import java.io.IOException;
import java.nio.file.Files;

public class RollbackHandlingManagerTest {
    private static final Context PPAPI_CONTEXT = ApplicationProvider.getApplicationContext();
    private static final String BASE_DIR = PPAPI_CONTEXT.getFilesDir().getAbsolutePath();

    private static final int DATASTORE_VERSION = 339900900;

    private BooleanFileDatastore mDatastore;

    @Before
    public void setup() {
        mDatastore =
                new BooleanFileDatastore(
                        PPAPI_CONTEXT.getFilesDir().getAbsolutePath(),
                        STORAGE_XML_IDENTIFIER,
                        DATASTORE_VERSION,
                        VERSION_KEY);
    }

    @After
    public void tearDown() {
        mDatastore.tearDownForTesting();
    }

    @Test
    public void testGetRollbackHandlingDataStoreDir() throws IOException {
        // The Datastore is in the directory with the following format.
        // /data/system/adservices/user_id/rollback/

        MockitoSession staticMockSession =
                ExtendedMockito.mockitoSession()
                        .spyStatic(Files.class)
                        .strictness(Strictness.LENIENT)
                        .startMocking();

        ExtendedMockito.doReturn(true).when(() -> Files.exists(any()));

        assertThat(
                        RollbackHandlingDatastoreLocationHelper
                                .getRollbackHandlingDataStoreDirAndCreateDir(
                                        /* baseDir */ "/data/system/adservices",
                                        /* userIdentifier */ 0))
                .isEqualTo("/data/system/adservices/0/rollback");

        assertThat(
                        RollbackHandlingDatastoreLocationHelper
                                .getRollbackHandlingDataStoreDirAndCreateDir(
                                        /* baseDir */ "/data/system/adservices",
                                        /* userIdentifier */ 1))
                .isEqualTo("/data/system/adservices/1/rollback");

        assertThrows(
                NullPointerException.class,
                () ->
                        RollbackHandlingDatastoreLocationHelper
                                .getRollbackHandlingDataStoreDirAndCreateDir(null, 0));

        staticMockSession.finishMocking();
    }

    @Test
    public void testGetOrCreateBooleanFileDatastore() throws IOException {
        RollbackHandlingManager rollbackHandlingManager =
                RollbackHandlingManager.createRollbackHandlingManager(
                        BASE_DIR, /* userIdentifier */ 0, DATASTORE_VERSION);
        BooleanFileDatastore datastore =
                rollbackHandlingManager.getOrCreateBooleanFileDatastore(
                        AdServicesManager.MEASUREMENT_DELETION);

        // Assert that the DataStore is created.
        assertThat(datastore).isNotNull();
    }

    @Test
    public void testRecordMeasurementDeletionOccurred() throws IOException {
        RollbackHandlingManager rollbackHandlingManager =
                RollbackHandlingManager.createRollbackHandlingManager(
                        BASE_DIR, /* userIdentifier */ 0, DATASTORE_VERSION);

        // By default, the bit is false.
        assertThat(
                        rollbackHandlingManager.wasAdServicesDataDeleted(
                                AdServicesManager.MEASUREMENT_DELETION))
                .isFalse();

        // Set the record bit and verify it is true.
        rollbackHandlingManager.recordAdServicesDataDeletion(
                AdServicesManager.MEASUREMENT_DELETION);
        assertThat(
                        rollbackHandlingManager.wasAdServicesDataDeleted(
                                AdServicesManager.MEASUREMENT_DELETION))
                .isTrue();

        // Reset the record bit and verify it is false.
        rollbackHandlingManager.resetAdServicesDataDeletion(AdServicesManager.MEASUREMENT_DELETION);
        assertThat(
                        rollbackHandlingManager.wasAdServicesDataDeleted(
                                AdServicesManager.MEASUREMENT_DELETION))
                .isFalse();
    }
}
